add timer to timeout the resize if it's too long
