+++ include/cwc/desktop/transaction.h
@@ -9,4 +9,6 @@ void transaction_schedule_tag(struct cwc_tag_info *tag);
 void transaction_pause();
 void transaction_resume();
 
+void transaction_commit(struct cwc_toplevel *toplevel);
+
 #endif // !_CWC_TRANSACTION_H
