+++ src/desktop/transaction.c
@@ -20,8 +20,12 @@
 #include <wayland-util.h>
 
 #include "cwc/desktop/layer_shell.h"
+#include "cwc/desktop/toplevel.h"
 #include "cwc/desktop/transaction.h"
+#include "cwc/layout/container.h"
 #include "cwc/server.h"
+#include "cwc/util.h"
+#include "private/container.h"
 
 static struct transaction {
     struct wl_event_source *idle_source;
@@ -137,4 +141,45 @@ void transaction_schedule_tag(struct cwc_tag_info *tag)
 void setup_transaction(struct cwc_server *s)
 {
     wl_array_init(&T.tags);
+}
+
+void transaction_commit(struct cwc_toplevel *toplevel)
+{
+    struct cwc_container *container = toplevel->container;
+
+    wlr_scene_node_set_position(&container->tree->node,
+                                container->pending.geom.x,
+                                container->pending.geom.y);
+    int gaps = cwc_container_get_gaps(container);
+    cwc_border_resize(&container->border,
+                      container->pending.geom.width - gaps * 2,
+                      container->pending.geom.height - gaps * 2);
+
+    if (wlr_box_empty(&toplevel->pending.clip)) {
+        wlr_scene_subsurface_tree_set_clip(&toplevel->surf_tree->node, NULL);
+    } else {
+        wlr_scene_subsurface_tree_set_clip(&toplevel->surf_tree->node,
+                                           &toplevel->pending.clip);
+    }
+
+    if (container->initializing && cwc_toplevel_is_visible(toplevel)) {
+        cwc_container_set_opacity(container, container->opacity_before);
+        container->initializing = false;
+    }
+
+    if (container->saved_tree) {
+        wlr_scene_node_destroy(&container->saved_tree->node);
+        container->saved_tree = NULL;
+        wlr_scene_node_set_enabled(&toplevel->surf_tree->node, true);
+    }
+
+    container->current      = container->pending;
+    toplevel->current       = toplevel->pending;
+    toplevel->pending       = (struct cwc_toplevel_state){0};
+    container->pending      = (struct cwc_container_state){0};
+    toplevel->resize_serial = 0;
+    toplevel->last_resize   = get_current_time_msec();
+
+    cwc_container_update_output(container);
+    wl_event_source_timer_update(container->resize_timer, 0);
 }
